Move socketio to lib
